- PseudospectralTimeseriesOutputComp moved to its own module within transcriptions.pseudospectral.components - PseudospectralTimeseriesOutputComp now detects when the output and input both use the same grid and output is to all nodes (same as input).  In this case, the interpolation matrix is Identity and the matrix multiplication is unnecessary. - When matrix multiplication IS necessary, it is done using a sparse CSR block matrix.  This improves performance more for larger and larger phases, with perhaps a slight penalty for small phases.
