+++ dymos/transcriptions/common/__init__.py
@@ -3,5 +3,6 @@ from .continuity_comp import RadauPSContinuityComp, GaussLobattoContinuityComp
 from .control_group import ControlGroup
 from .polynomial_control_group import PolynomialControlGroup
 from .path_constraint_comp import PathConstraintComp
-from .timeseries_output_comp import PseudospectralTimeseriesOutputComp
+from ..pseudospectral.components.pseudospectral_timeseries_output_comp import \
+    PseudospectralTimeseriesOutputComp
 from .time_comp import TimeComp
