+++ dymos/transcriptions/pseudospectral/components/__init__.py
@@ -3,3 +3,4 @@ from .state_interp_comp import StateInterpComp
 from .state_independents import StateIndependentsComp
 from .control_endpoint_defect_comp import ControlEndpointDefectComp
 from .gauss_lobatto_interleave_comp import GaussLobattoInterleaveComp
+from .pseudospectral_timeseries_output_comp import PseudospectralTimeseriesOutputComp
